Add definitions for Dell U4919DW. (#197)
This monitor presents itself under various EDID identifiers, depending on
which video port is currently used, and sometimes whether PBP is enabled. It's
also likely that HDMI1.4 vs HDMI2.0 makes a difference, but I don't have
HDMI2.0 compliant hardware to test this.

The main definitions are in U4919DW.xml, and all the other valid EDID
identifiers include that file.

* `DELA107`: using HDMI-1
* `DELA109`: using HDMI-2
* `DELA10A`: using HDMI-1 in PBP
* `DELA10B`: using HDMI-2 in PBP
* `DELA10D` and `DELA10E`: using DisplayPort. I could not reliably determine
  which (if any) is for PBP, getting both identifiers whether or not using PBP.
* `DELA10F`: using USB-C
* `DELA111`: using USB-c in PBP

Some definitions cribbed from other monitors, especially U2719D (DEL415A).

A new control is E7, which is related to the monitor's KVM interface. It has two
USB inputs, which can be assigned in the OSD to any of its video ports. Using E7
seems to assign the specific USB port to a video port, regardless if it's
currently selected, or visible in PBP mode. I was only able to discover a couple
of settings; there may be more logic to the value of this control.